requirements.txt
google-generativeai
streamlit
streamlit-option-menu
pillow